GEOMESA-877 Move SFTBuilder to utils from accumulo
Moved SftBuilder to utils and made AccumuloSftBuilder, which extends SftBuilder, to get access to splitter information. Refactored classes using SftBuilder to import the correct class.
